major changes
- refactor code + structure
- add unit tests
